Add Flatpak packaging, fix compiler warnings, and improve code clarity
- Add Flatpak manifest and desktop/metadata files for Linux distribution
- Update CI workflow to use self-hosted runners and build Flatpak releases
- Fix compiler warnings: add WIN32_LEAN_AND_MEAN/NOMINMAX guards, cast unused return values
- Rename variables for clarity: player→caster in action_queue, combat_y→cy in icon_panel
- Add explicit casts for camera position to int32_t in world rendering

+++ src/world/world.cpp
@@ -64,7 +64,7 @@ void world::render(renderer& rend)
     // Update map renderer with current zoom level for proper tile culling
     map_renderer_.set_zoom_level(zoom_mode_enabled_ ? static_cast<float>(zoom_level_) : 1.0f);
 
-    map_renderer_.render(rend, current_map_, camera_x_, camera_y_);
+    map_renderer_.render(rend, current_map_, static_cast<int32_t>(camera_x_), static_cast<int32_t>(camera_y_));
 
     // Weather particles and day/night overlay are rendered by weather_system
     // in game_state_manager::render_playing() after the scene is drawn
@@ -73,17 +73,17 @@ void world::render(renderer& rend)
 void world::render_terrain(renderer& rend)
 {
     map_renderer_.set_zoom_level(zoom_mode_enabled_ ? static_cast<float>(zoom_level_) : 1.0f);
-    map_renderer_.render_terrain(rend, current_map_, camera_x_, camera_y_);
+    map_renderer_.render_terrain(rend, current_map_, static_cast<int32_t>(camera_x_), static_cast<int32_t>(camera_y_));
 }
 
 void world::render_objects_row(renderer& rend, int32_t row_y, const sf::IntRect* player_bounds)
 {
-    map_renderer_.render_objects_row(rend, current_map_, row_y, camera_x_, camera_y_, player_bounds);
+    map_renderer_.render_objects_row(rend, current_map_, row_y, static_cast<int32_t>(camera_x_), static_cast<int32_t>(camera_y_), player_bounds);
 }
 
 map_renderer::visible_range world::get_visible_tile_range() const
 {
-    return map_renderer_.get_visible_range(current_map_, camera_x_, camera_y_);
+    return map_renderer_.get_visible_range(current_map_, static_cast<int32_t>(camera_x_), static_cast<int32_t>(camera_y_));
 }
 
 void world::apply_zoom_view(renderer& rend)
